字幕转化时确保字幕文件存在

+++ cmd/srt2vtt.js
@@ -21,13 +21,17 @@ function convertSrt2vtt(r){
     srtPath = path.resolve(r, './subtitle');
     var srt2vtt = require('srt2vtt');
 
-    var srtData = fs.readFileSync(srtPath);
-    srt2vtt(srtData, function(err, vttData) {
-        if (err) throw new Error(err);
-            
-        var storePos = path.resolve(r, 'subtitle.vtt')
-        fs.writeFileSync(storePos, vttData);
-    });
+    fs.exists(srtPath, function(doExist){
+        if(doExist){
+            var srtData = fs.readFileSync(srtPath);
+            srt2vtt(srtData, function(err, vttData) {
+                if (err) throw new Error(err);
+                    
+                var storePos = path.resolve(r, 'subtitle.vtt')
+                fs.writeFileSync(storePos, vttData);
+            });
+        }
+    })
 }
 
 function copyM3u8(storeRoot){
